Update: improve response time for sending invite by using go routines

+++ server/handlers/team_controllers/delete_team.go
@@ -12,10 +12,17 @@ func DeleteTeamController(ctx *fiber.Ctx) error {
 	user := ctx.Locals("user").(*models.UserRes)
 	teamId := utils.HexToObjectID(ctx.Params("id"))
 
+	if teamExist := models.CheckTeamExist(teamId, ctx.Context()); !teamExist {
+		return ctx.Status(404).JSON(fiber.Map{
+			"success": false,
+			"message": "Team does not exist",
+		})
+	}
+
 	if err := models.DeleteTeam(ctx.Context(), teamId, user.ID); err != nil {
 		return ctx.Status(500).JSON(fiber.Map{
 			"success": false,
-			"message": "Failed to delete team: " + err.Error(),
+			"message": err.Error(),
 		})
 	}
